fastapi_service/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
